% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cbCheck}
\alias{cbCheck}
\title{Colorblindness check}
\usage{
cbCheck(pal)
}
\arguments{
\item{pal}{color palette to take}
}
\value{
returns a plot of different color blindness check
}
\description{
Colorblindness check
}
