% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{clashCheck}
\alias{clashCheck}
\title{Clash check}
\usage{
clashCheck(pal)
}
\arguments{
\item{palette}{Name of a palette to check

 @return A distance between palette components
 @export 
 @importFrom grDevices col2rgb}
}
\description{
Clash check
}
