% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{clashCheckWhite}
\alias{clashCheckWhite}
\title{Clash check with white}
\usage{
clashCheckWhite(pal)
}
\arguments{
\item{palette}{Name of a palette to check

 @return A distance between palette components
 @export}
}
\description{
Clash check with white
}
