% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{col2hsv}
\alias{col2hsv}
\title{Convert HEX values to HSV values}
\usage{
col2hsv(HEXVEC)
}
\arguments{
\item{HEXVEC}{hex code string vector}
}
\value{
a dataframe of HSV values with columns "hex", "h", "s", "v"
}
\description{
Convert HEX values to HSV values
}
