% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{col2rgb2}
\alias{col2rgb2}
\title{Convert HEX codes to RGB data frame}
\usage{
col2rgb2(hexvec)
}
\arguments{
\item{hexvec}{a vector of HEX codes}
}
\value{
a dataframe with columns "r", "g", "b"
}
\description{
Convert HEX codes to RGB data frame
}
