% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{createDivergent}
\alias{createDivergent}
\title{Create a divergent color palette}
\usage{
createDivergent(
  num_cols = 9,
  color = NULL,
  plot = TRUE,
  interpolation = "spline"
)
}
\arguments{
\item{num_cols}{optional - is the user defined number of colors in the desired palette}

\item{color}{optional - user defined color. This can be NULL, a HEX code or a R-defined color}

\item{plot}{TRUE/FALSE whether to display the palette or not}

\item{interpolation}{which method to use to interpolate. This can be either a spline or linear

 @return HEX codes of desired palette and (optional) a plot
 @export}
}
\description{
Create a divergent color palette
}
