% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{createTriad}
\alias{createTriad}
\title{Create a triad color palette}
\usage{
createTriad(num_cols = 5, color = NULL, plot = TRUE, interpolation = "spline")
}
\arguments{
\item{num_cols}{optional - is the user defined number of colors in the desired palette}

\item{color}{optional - user defined color. This can be NULL, a HEX code or a R-defined color}

\item{plot}{TRUE/FALSE whether to display the palette or not}

\item{interpolation}{which method to use to interpolate. This can be either a spline or linear

 @return HEX codes of desired palette and (optional) a plot
 @export}
}
\description{
Create a triad color palette
}
