% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{displayPals}
\alias{displayPals}
\title{Palette Visualization}
\usage{
displayPals(palette = NULL)
}
\arguments{
\item{palette}{Name of a palette to view. Default is NULL which will display all the built-in palettes

 @return Plots}
}
\description{
Palette Visualization
}
