% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getHue}
\alias{getHue}
\title{get Hue}
\usage{
getHue(color = NULL)
}
\arguments{
\item{color}{optional - user defined color. This can be NULL, a HEX code or a R-defined color

 @return a valid hue 
 @export}
}
\description{
get Hue
}
