% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{lighten}
\alias{lighten}
\title{Increase value of a palette}
\usage{
lighten(pal, plot = T)
}
\arguments{
\item{pal}{String of hex codes}

\item{plot}{TRUE/FALSE display plot of new palette

 @return Lightened color palette 
 @export}
}
\description{
Increase value of a palette
}
