% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotCols}
\alias{plotCols}
\title{Plotting fuction}
\usage{
plotCols(hex = NULL, hsv_df = NULL)
}
\arguments{
\item{hex}{HEX codes to plot}

\item{hsv_df}{HSV dataframe to plot - only one of HEX or HSV can be non-null}
}
\value{
plotted colors returned
}
\description{
Plotting fuction
}
