% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{posNegCheck}
\alias{posNegCheck}
\title{Checks to ensure hsv code relates to valid color}
\usage{
posNegCheck(hsvdf)
}
\arguments{
\item{hsvdf}{a data frame or matrix with columns h, s, v

 @return a df or matrix that specifies a valid color
 @export}
}
\description{
Checks to ensure hsv code relates to valid color
}
