% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ranColorGenerator}
\alias{ranColorGenerator}
\title{Random palette generator}
\usage{
ranColorGenerator(num_cols)
}
\arguments{
\item{num_cols}{Number of colors in returned palette}
}
\value{
Random palette returned (and plotted)
}
\description{
Random palette generator
}
