% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{relLumRatio}
\alias{relLumRatio}
\title{Relative Luminance ratio}
\usage{
relLumRatio(pal)
}
\arguments{
\item{palette}{Name of a palette to check

 @return A ratio
 @export}
}
\description{
Relative Luminance ratio
}
