% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{reversePal}
\alias{reversePal}
\title{Reverses a palette}
\usage{
reversePal(hexvec)
}
\arguments{
\item{hexvec}{hex code string vector}
}
\value{
hex code vector but like, reversed
}
\description{
Reverses a palette
}
