% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rgb2hsv2}
\alias{rgb2hsv2}
\title{Convert RGB values to HSV values}
\usage{
rgb2hsv2(df)
}
\arguments{
\item{df}{a dataframe of RGB values with columns "r", "g", "b"}
}
\value{
a dataframe of HSV values with columns "h", "s", "v"
}
\description{
Convert RGB values to HSV values
}
