% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scale_color_colorway_d}
\alias{scale_color_colorway_d}
\title{Color discrete data on a ggplot}
\usage{
scale_color_colorway_d(
  palette,
  na.value = "grey50",
  limits = NULL,
  reverse = FALSE,
  user_cols = NULL,
  num_cols = 5
)
}
\arguments{
\item{palette}{Palette name from default palettes or your created palette stored as a hex code vector}

\item{na.value}{value for nas in data, default is middle grey}

\item{limits}{numeric vector of length 2 providing limits of scale. NULL defaults to scale range}

\item{reverse}{TRUE/FALSE on whether to reverse the color palette or not}

\item{user_cols}{user specified colors for the palette}

\item{num_cols}{should correspond to number of categories 

 @return Function to be used for filling ggplot continuous data with chosen palette
 @export
 @import ggplot2}
}
\description{
Color discrete data on a ggplot
}
