% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scale_fill_colorway_c}
\alias{scale_fill_colorway_c}
\title{Fill continuous data on a ggplot}
\usage{
scale_fill_colorway_c(
  palette,
  na.value = "grey50",
  limits = NULL,
  reverse = FALSE,
  user_cols = NULL
)
}
\arguments{
\item{palette}{Palette name from default palettes or your created palette stored as a hex code vector}

\item{na.value}{value for nas in data, default is middle grey}

\item{limits}{numeric vector of length 2 providing limits of scale. NULL defaults to scale range}

\item{reverse}{TRUE/FALSE on whether to reverse the color palette or not}

\item{user_cols}{user specified colors for the palette

 @return Function to be used for filling ggplot continuous data with chosen palette
 @import ggplot2
 @export}
}
\description{
Fill continuous data on a ggplot
}
