% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{shadePicker}
\alias{shadePicker}
\title{Pick a shade}
\usage{
shadePicker(hex1)
}
\arguments{
\item{hex1}{the HEX code of the color to be replaced}
}
\value{
A list of HEX codes representing color replacements and a plot
}
\description{
If you want to replace a color in your color way with another, shadePicker
helps you find a new shade
}
