% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stretchPal}
\alias{stretchPal}
\title{Interpolate - stretch out a discrete palette}
\usage{
stretchPal(user_cols, num_cols, interpolation = "linear")
}
\arguments{
\item{user_cols}{Palette to be stretched - can be R defined colors or HEX codes}

\item{num_cols}{Number of colors in returned palette}
}
\value{
A vector of hex codes that represent that 'stretched' out discrete palette
}
\description{
Interpolate - stretch out a discrete palette
}
