% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{structurePal}
\alias{structurePal}
\title{Visualize palette  structure}
\usage{
structurePal(pal)
}
\arguments{
\item{palette}{Name of a palette to view. 

 @return Plots  
 @export
 @import ggplot2}
}
\description{
Visualize palette  structure
}
